frontend/requirements.txt
flask>=3.0.6,<4
python-dotenv>=1.0.0,<2
requests>=2.31.0,<3
